+++ src/SWP/Bundle/ContentBundle/Resources/config/doctrine-orm/ArticleSeoMetadata.orm.yml
@@ -4,18 +4,21 @@ SWP\Bundle\ContentBundle\Model\ArticleSeoMetadata:
   oneToOne:
       metaMedia:
         targetEntity: SWP\Bundle\ContentBundle\Model\ArticleSeoMediaInterface
+        cascade: ["persist"]
         joinColumn:
           name: seo_meta_media_id
           referencedColumnName: id
           nullable: true
       ogMedia:
         targetEntity: SWP\Bundle\ContentBundle\Model\ArticleSeoMediaInterface
+        cascade: ["persist"]
         joinColumn:
           name: seo_og_media_id
           referencedColumnName: id
           nullable: true
       twitterMedia:
         targetEntity: SWP\Bundle\ContentBundle\Model\ArticleSeoMediaInterface
+        cascade: ["persist"]
         joinColumn:
           name: seo_twitter_media_id
           referencedColumnName: id
